introduce a drain router
If you echo another host to the DRAIN_FILE

echo smtp2.immerda.ch > /etc/exim/drain_to_host

and restart queue processing, this host will forward any emails
still in the queue (and any incoming mails) to the host you want
to drain the emails to.
This makes it possible to drain the queue to another host before
taking it offline for maintenance or decommisioning.